+++ src/engineio/client/mod.rs
@@ -1,0 +1,2 @@
+mod socket;
+pub use {socket::Socket, socket::SocketBuilder};
