+++ src/engineio/packet.rs
@@ -90,8 +90,11 @@ impl TryFrom<Packet> for HandshakePacket {
 
 impl Packet {
     /// Creates a new `Packet`.
-    pub fn new(packet_id: PacketId, data: Bytes) -> Self {
-        Packet { packet_id, data }
+    pub fn new<T: Into<Bytes>>(packet_id: PacketId, data: T) -> Self {
+        Packet {
+            packet_id,
+            data: data.into(),
+        }
     }
 }
 
@@ -145,7 +148,8 @@ impl From<Packet> for Bytes {
     }
 }
 
-pub struct Payload(Vec<Packet>);
+#[derive(Debug, Clone)]
+pub(crate) struct Payload(Vec<Packet>);
 
 impl Payload {
     // see https://en.wikipedia.org/wiki/Delimiter#ASCII_delimited_text
@@ -157,6 +161,12 @@ impl Payload {
 
     pub fn len(&self) -> usize {
         self.0.len()
+    }
+
+    pub fn iter(&self) -> Iter {
+        Iter {
+            iter: self.0.iter(),
+        }
     }
 }
 
@@ -197,6 +207,17 @@ impl TryFrom<Payload> for Bytes {
         // remove the last separator
         let _ = buf.split_off(buf.len() - 1);
         Ok(buf.freeze())
+    }
+}
+
+pub struct Iter<'a> {
+    iter: std::slice::Iter<'a, Packet>,
+}
+
+impl<'a> Iterator for Iter<'a> {
+    type Item = &'a Packet;
+    fn next(&mut self) -> std::option::Option<<Self as std::iter::Iterator>::Item> {
+        self.iter.next()
     }
 }
