+++ src/engineio/transport.rs
@@ -1,4 +1,5 @@
 use super::transports::{PollingTransport, WebsocketSecureTransport, WebsocketTransport};
+use crate::engineio::packet::{Packet, Payload};
 use crate::error::Result;
 use adler32::adler32;
 use bytes::Bytes;
@@ -72,3 +73,5 @@ impl std::fmt::Debug for dyn Transport {
         f.write_fmt(format_args!("Transport(base_url: {:?})", self.base_url(),))
     }
 }
+
+//TODO: add iter() implementation
