+++ src/socketio/mod.rs
@@ -1,11 +1,11 @@
+pub mod client;
 /// Defines the events that could be sent or received.
 pub mod event;
 mod packet;
 /// Defines the types of payload (binary or string), that
 /// could be sent or received.
 pub mod payload;
-pub mod socket;
-pub(crate) mod transport;
+pub(self) mod transport;
 
 #[cfg(test)]
 pub(crate) mod test {
